requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
ecdsa==0.19.1
fastapi==0.134.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruff==0.15.4
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
